% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{BasisPolynomial.raw}
\alias{BasisPolynomial.raw}
\title{Bernstein polynomial}
\usage{
BasisPolynomial.raw(t, psi, mint, maxt)
}
\arguments{
\item{t}{Values at which the Bernstein polynomial are evaluated}

\item{psi}{Vector of coefficients for the Bernstein basis polynomials. The dimension of the vector is equal to the degree plus 1}

\item{mint}{Lower boundary of the time interval}

\item{maxt}{Upper boundary of the time interval}
}
\value{
Values of the Bernstein polynomial at t.
}
\description{
Evaluate the Bernstein polynomial at given time points using user-provided coefficients and support.
}
\details{
The degree of Bernstein polynomial is m. \eqn{\Psi}  is an (m+1)-dimentsional vector of coefficients of the basis polynomials. If \eqn{\Psi} does not be input, the values of psi will be calculated based on the parameter \eqn{\lambda}.
}
\examples{
t <- seq(1,10,1)
psi <- seq(0,1,0.25)
BasisPolynomial.raw(t = t, psi = psi, mint = 0, maxt = 5)

}
