% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{BernsteinPolynomial}
\alias{BernsteinPolynomial}
\title{Bernstein polynomial}
\usage{
BernsteinPolynomial(t, psi, mint, maxt)
}
\arguments{
\item{t}{Values at which the Bernstein polynomial are evaluated}

\item{psi}{Vector of coefficients for the Bernstein polynomials. The dimension of the vector is equal to the degree plus 1}

\item{mint}{Lower boundary of the time interval}

\item{maxt}{Upper boundary of the time interval}
}
\value{
A vector of values of the Bernstein polynomial at t.
}
\description{
Evaluate the Bernstein polynomial at given time points using user-provided coefficients and support.
}
\details{
The degree of Bernstein polynomial is \code{m}. \code{psi} is an (m+1)-dimentsional vector of coefficients of the Bernstein polynomials. Then, this function calculate the values of the Bernstein polynomial at t.
}
\examples{
t <- seq(1,10,1)
psi <- seq(0.25,1,0.25)
BernsteinPolynomial(t = t, psi = psi, mint = 0, maxt = 5)
#0.2 0.4 0.6 0.8 1.0 1.2 1.4 1.6 1.8 2.0

}
