% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ICDASim}
\alias{ICDASim}
\title{Data simulator}
\usage{
ICDASim(seed = NA, n, beta00, beta10, beta20, rho, gamma, cs)
}
\arguments{
\item{seed}{Seed of the random generator (optional)}

\item{n}{Cluster index}

\item{beta00}{True parameter value of the first component of \eqn{\beta}}

\item{beta10}{True parameter value of the second component of \eqn{\beta}}

\item{beta20}{True parameter value of the third component of \eqn{\beta}}

\item{rho}{Transformation parameter \eqn{\rho} in the Box-Cox transformation}

\item{gamma}{Parameter \eqn{\gamma} in in the Gumbel Copula}

\item{cs}{Maximum cluster size}
}
\value{
A data frame consisting of the following columns: \itemize{
  \item \strong{family} : Cluster index
  \item \strong{Ci} : Observation time
  \item \strong{delta} : Event indicator
  \item \strong{x1} : 1st covariate
  \item \strong{x2} : 2nd covariate
}
}
\description{
Generate a data set based on the simulation setting in Lam et al. (2021) <DOI: \href{https://doi.org/10.1002/sim.8910}{10.1002/sim.8910}>
}
\details{
See Section 4 of Lam et al. (2021) <DOI: \href{https://doi.org/10.1002/sim.8910}{10.1002/sim.8910}> for the data generation mechanism.
}
\examples{
data <- ICDASim(seed = 1942, n = 100, beta00 = 0.5, beta10 = -1, beta20 = 1,
cs = 40, rho = 0.5,gamma = 0.5)
## Generate 100 clusters with maximum size 40
}
