% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{BernsteinPolynomial}
\alias{BernsteinPolynomial}
\title{Evaluate a user-provided Bernstein polynomial}
\usage{
BernsteinPolynomial(t, psi, mint, maxt)
}
\arguments{
\item{t}{Values at which the Bernstein polynomial is evaluated}

\item{psi}{Vector of coefficients for the Bernstein polynomial (not including the intercept). The dimension of the vector is equal to the degree of the polynomial. The intercept is assumed to be 0}

\item{mint}{Lower boundary of the support of the polynomial}

\item{maxt}{Upper boundary of the support of the polynomial}
}
\value{
A vector of values of the Bernstein polynomial at \code{t}.
}
\description{
Evaluate the Bernstein polynomial at given time points using user-provided coefficients and support.
}
\details{
The vector \code{psi} contains the coefficients of the Bernstein basis polynomial. The polynomial is supported on ( \code{mint} , \code{maxt} ) and is equal to zero at \code{mint}. The intercept should not be included in \code{psi}. The function returns the values of the polynomial at the time points in \code{t}.
}
\examples{
t <- seq(1,10,1)
psi <- seq(0.25,1,0.25)
BernsteinPolynomial(t = t, psi = psi, mint = 0, maxt = 5)
#0.2 0.4 0.6 0.8 1.0 1.2 1.4 1.6 1.8 2.0

}
