% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Est.ICScure}
\alias{Est.ICScure}
\title{Estimate model parameters and standard errors}
\usage{
Est.ICScure(
  data,
  rho,
  degree,
  weighting = TRUE,
  t.min = 0,
  t.max = NA,
  reltol = 10^-6,
  maxit = 1000,
  Hessian = TRUE
)
}
\arguments{
\item{data}{A matrix with n rows, where n is the number of observations. The first column is the nth cluster, the second column is event time of interest, The third column is event indicator \eqn{\delta}. The forth column to the end are the covariates (\eqn{x_1,x_2,...,x_n})}

\item{rho}{Transformation parameter \eqn{\rho}}

\item{degree}{Degree of Bernstein polynomial}

\item{weighting}{Logical, if TRUE, then CWGEE is adopted; otherwise, GEE is adopted. Default is TRUE}

\item{t.min}{Left end point of the support of the distribution function F}

\item{t.max}{Cure threshold. If NA, then it is set to be the largest inspection time with \eqn{\Delta} = 1.}

\item{reltol}{Relative tolerance to input to the optimization function of R from "stats" package}

\item{maxit}{Maximum number of iterations}

\item{Hessian}{Logical; If TRUE, then the standard error of the regression parameter estimators would be estimated}
}
\value{
The followings components will be returned: \itemize{
  \item \strong{degree} : Degree of Bernstein Polynomial
  \item \strong{psi} : Vector of the coefficients of the Bernstein basis polynomial in the distribution function F
  \item \strong{beta} : Vector of the regression parameters \eqn{\beta}
  \item \strong{betaSE} : Estimated standard error of the estimator of \eqn{\beta}
  \item \strong{iteration} : Numbers of iteration used in the optimization process
  \item \strong{covergence} : Indicator of the convergence of the optimization
  \item \strong{AIC} : Akaike information criterion (AIC)
  \item \strong{ct} : Cure Threshold
}
}
\description{
Perform the cluster-weighted GEE or GEE estimation of Lam et al. (2021) <DOI: \href{https://doi.org/10.1002/sim.8910}{10.1002/sim.8910}>
}
\examples{
Dataset <- ICDASim(seed = 1942, N = 100, beta00 = 0.5, beta10 = -1, beta20 = 1,
cs = 40, rho=0.5,PoAlpha = 0.5, zeta=0.75)
Result <- Est.ICScure(data = Dataset, rho = 0.5, degree = 3, weighting = FALSE)

}
