% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Est.ICScure}
\alias{Est.ICScure}
\title{Estimate model parameters and standard errors}
\usage{
Est.ICScure(
  data,
  rho,
  degree,
  weighting = TRUE,
  t.min = 0,
  t.max = NA,
  reltol = 10^-6,
  maxit = 1000,
  Hessian = TRUE
)
}
\arguments{
\item{data}{A \code{n x (p+3)} matrix, where \code{n} is the sample size and \code{p} is the number of covariates. The first column consists of cluster indices, the second column consist of the observation time, the third column consist of the event indicator, and the forth to the last columns consist of the covariates (not including the intercept). The set of covariates can be empty.}

\item{rho}{Transformation parameter \eqn{\rho} in the Box-Cox transformation}

\item{degree}{Degree of the Bernstein polynomial}

\item{weighting}{Logical; If TRUE, then CWGEE is adopted; otherwise, GEE is adopted. Default is TRUE}

\item{t.min}{Left end point of the support of the distribution function \emph{\bold{F}}}

\item{t.max}{Cure threshold. If \code{NA}, then it is set to be the largest observation with \eqn{\Delta = 1}}

\item{reltol}{Relative tolerance for the function \code{optim} of the \code{stat} package}

\item{maxit}{Maximum number of iterations for the function \code{optim} of the \code{stat} package}

\item{Hessian}{Logical; If TRUE, then the standard error of the regression parameter estimators would be estimated}
}
\value{
A list of the following components: \itemize{
  \item \strong{degree} : Degree of the Bernstein polynomial
  \item \strong{psi} : Vector of the coefficients of the Bernstein basis polynomial in the estimated distribution function \emph{\bold{F}}
  \item \strong{beta} : Vector of the regression parameters \eqn{\beta}
  \item \strong{betaSE} : Estimated standard error of the estimator for \eqn{\beta}
  \item \strong{iteration} : Numbers of iteration used the function \code{optim}
  \item \strong{covergence} : Logical; If TRUE, then the optimization algorithm is converged
  \item \strong{AIC} : Akaike information criterion (AIC) of the fitted model
}
}
\description{
Perform the cluster-weighted GEE or GEE estimation in Lam et al. (2021) <DOI: \href{https://doi.org/10.1002/sim.8910}{10.1002/sim.8910}>
}
\examples{
Dataset <- ICDASim(seed = 1942, n = 100, beta00 = 0.5, beta10 = -1, beta20 = 1,
cs = 40, rho=0.5,gamma = 0.5)
Result <- Est.ICScure(data = Dataset, rho = 0.5, degree = 3, weighting = FALSE)
}
\references{
Lam, K. F., Lee, C. Y., Wong, K. Y., & Bandyopadhyay, D. (2021). Marginal analysis of current status data with informative cluster size using a class of semiparametric transformation cure models. Statistics in Medicine [online], DOI:10.1002/sim.8910
}
\seealso{
\code{BernsteinPolynomial}
}
\author{
Wong Kin Yau, Alex <kin-yau.wong@polyu.edu.hk>
}
