% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{ICDASim}
\alias{ICDASim}
\title{Data simulator}
\usage{
ICDASim(seed = NA, N, beta00, beta10, beta20, rho, PoAlpha, cs, zeta)
}
\arguments{
\item{seed}{Seed of the random generator (optional)}

\item{N}{Number of clusters}

\item{beta00}{True parameter value for beta0}

\item{beta10}{True parameter value for beta1}

\item{beta20}{True parameter value for beta2}

\item{rho}{Transformation parameter \eqn{\rho}}

\item{PoAlpha}{Parameter in the Gumbel Copula}

\item{cs}{Maximum cluster size}

\item{zeta}{Parameter in the binomial distribution}
}
\value{
The followings components will be returned: \itemize{
  \item \strong{family} : The nth cluster (subject)
  \item \strong{Ci} : Event time of interest
  \item \strong{delta} : Event indicator
  \item \strong{x1} : 1st covariates
  \item \strong{x2} : 2nd covariates
}
}
\description{
Generate a data set according to the simulation studies in Lam et al. (2021) <DOI: \href{https://doi.org/10.1002/sim.8910}{10.1002/sim.8910}>
}
\details{
From the simulator, two covariates are generated, namely X1 from a Bernoulli distribution with probability 0.5 and X2 from a standard normal distribution.
The cluster size N is randomly drawn from a binomial distribution based on the latent variable \eqn{\xi}.
Then, the event time for the ith cluster can be generated based on the Gumbel copula model with the join survival function.
}
\examples{
data<-ICDASim(seed = 1942, N = 100, beta00 = 0.5, beta10 = -1, beta20 = 1,
cs = 40, rho = 0.5,PoAlpha = 0.5, zeta = 0.75)
## Generate 100 clusters with maximum size 40
}
